+++ Thunderstore/Package Files/CHANGELOG.md
@@ -1,8 +1,7 @@
 ## 0.10.0
 
 Item highlights being survivor colored now use a better method
-- Should be more performant at high item counts, if you still notice any lag while this is enabled let me know.
-- Because of (hopefully) better performance the config option for this is now on by default.
+- Should be more performant at high item counts so the config option for this is now on by default.
 
 Changed item highlight texture to a cleaner one
 
@@ -20,6 +19,8 @@ Made HP bar text smoother
 Changed "Lv: ##" text to "Level ##"
 
 Fixed NRE when opening inventories menu when someone had no items
+
+Fixed consistent difficulty bar coloring not working past the start of "HAHAHAHA"
 
 ## 0.9.0